Add cross-API pipeline layout abstraction and support
Introduce PipelineLayout handle, PushConstant, PushDescriptor, and PipelineLayoutDesc structs to Wisdom API. Add ShaderStages and DescriptorType enums. Implement wisDeviceCreatePipelineLayout and wisDestroyPipelineLayout for both DX12 and Vulkan, with full backend support for push constants and push descriptors. Update codegen, documentation, and XML registry. Improve C++ wrapper codegen and Vulkan extension buffer handling. This enables advanced, flexible resource binding across APIs.

+++ src/include/wisdom/impl/vulkan/vk_extensions.hpp
@@ -177,11 +177,11 @@ public:
     WIS_INLINE const VkExtensionProperties* GetExtensionProperties(const char* name) const noexcept;
 
     struct InitBuffer {
-        std::unique_ptr<uint8_t[]> buffer;
-        const char**               extension_names  = nullptr;
-        std::size_t                extension_count  = 0;
-        VkBaseOutStructure*        feature_structs  = nullptr;
-        VkBaseOutStructure*        property_structs = nullptr;
+        std::unique_ptr<uint64_t[]> buffer;
+        const char**                extension_names  = nullptr;
+        std::size_t                 extension_count  = 0;
+        VkBaseOutStructure*         feature_structs  = nullptr;
+        VkBaseOutStructure*         property_structs = nullptr;
     };
     WIS_INLINE InitBuffer GetInitBuffer(WisResult& out_res) const noexcept;
